html/rendering/replaced-elements/attributes-for-embedded-content-and-images/embedded-and-images-presentational-hints-ascii-case-insensitive.html

<!DOCTYPE html>
<meta charset="utf-8">
<link rel="help" href="https://html.spec.whatwg.org/#attributes-for-embedded-content-and-images:presentational-hints">
<link rel="help" href="https://drafts.csswg.org/selectors-4/#attribute-case">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<img src="fuchsia.png" align="absbottom">
<img src="fuchsia.png" align="AbSbOtToM">
<img src="fuchsia.png" align="abſbottom">
<img src="fuchsia.png" align="abscenter">
<img src="fuchsia.png" align="AbScEnTeR">
<img src="fuchsia.png" align="abſcenter">
<img src="fuchsia.png" align="absmiddle">
<img src="fuchsia.png" align="AbSmIdDlE">
<img src="fuchsia.png" align="abſmiddle">
<script>
test(() => {
  const img = document.querySelectorAll("img");

  assert_equals(getComputedStyle(img[0]).getPropertyValue("vertical-align"), "bottom");
  assert_equals(getComputedStyle(img[1]).getPropertyValue("vertical-align"), "bottom");
  assert_equals(getComputedStyle(img[2]).getPropertyValue("vertical-align"), "baseline");
  assert_equals(getComputedStyle(img[3]).getPropertyValue("vertical-align"), "middle");
  assert_equals(getComputedStyle(img[4]).getPropertyValue("vertical-align"), "middle");
  assert_equals(getComputedStyle(img[5]).getPropertyValue("vertical-align"), "baseline");
  assert_equals(getComputedStyle(img[6]).getPropertyValue("vertical-align"), "middle");
  assert_equals(getComputedStyle(img[7]).getPropertyValue("vertical-align"), "middle");
  assert_equals(getComputedStyle(img[8]).getPropertyValue("vertical-align"), "baseline");
});
</script>
